change contracts lib structure for better readability
Backend:
- fix compilation by excluding "seed" folder

Other:
- updated @types/node in all workspaces so that all @ts-rest/core resolves to the same pnpm dependency with the same types file
