+++ README.md
@@ -32,13 +32,13 @@ pnpm dev
 
 ## TODO
 
-- [] products and filters
-- [] featured products
-- [] search by products
-- [] cart
-- [] payment and orders
-- [] favorites
-- [] rating and comments
-- [] simple analytics in dashboard
-- [] improve build process
-- [] deploy
+- [ ] products and filters
+- [ ] featured products
+- [ ] search by products
+- [ ] cart
+- [ ] payment and orders
+- [ ] favorites
+- [ ] rating and comments
+- [ ] simple analytics in dashboard
+- [ ] improve build process
+- [ ] deploy
