+++ apps/backend/package.json
@@ -68,7 +68,7 @@
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.2",
     "@types/morgan": "^1.9.9",
-    "@types/node": "^20.3.1",
+    "@types/node": "^22.10.1",
     "@types/pg": "^8.11.10",
     "@types/supertest": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^8.12.0",
