+++ apps/web-store/package.json
@@ -42,7 +42,7 @@
     "@stylistic/eslint-plugin": "^2.9.0",
     "@total-typescript/ts-reset": "^0.6.1",
     "@types/eslint__js": "^8.42.3",
-    "@types/node": "^20",
+    "@types/node": "^22.10.1",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@typescript-eslint/parser": "^8.0.0",
