+++ libs/contracts/src/dto/color.dto.ts
