+++ libs/contracts/src/dto/session.dto.ts
